+++ telegram/src/main/kotlin/me/ivmg/telegram/network/ApiClient.kt
@@ -82,7 +82,7 @@ class ApiClient(
         parseMode: String?,
         disableWebPagePreview: Boolean?,
         disableNotification: Boolean?,
-        replyToMessageId: Int?,
+        replyToMessageId: Long?,
         replyMarkup: ReplyMarkup?
     ): Call<Response<Message>> {
